requirements.txt
streamlit
torch
diffusers
transformers
pillow